Refactor generator utility: Extract generator decorators and classes
- Move `consumer` and `compose` decorators, `GenState` and `Gen` classes to `ezmsg.util.generator`.
- Clean up `ezmsg_generator.py` by removing the above elements.
- Add docstring to `gen_to_unit.py` explaining its experimental status.
